Nextflow DSL 1 pipeline for ttv primer analysis
This pipeline focuses on finding three primers in the untranslated regions (UTR) that are commonly used for PCR analysis of Torque teno virus (TTV). It finds primers AMTS UTR, AMTAS UTR and AMTPTU UTR. Primers found in sequences are based on exact matching, one mismatch and two mismatches. Outputs are calculations of sequences and tables with summaries of sequences that contain at least one primer. Different bar charts are also created based on the results from the tables.